constant 분리
