+++ src/Constants/common.ts
@@ -1,0 +1,1 @@
+export const BASE_URL_TEST = 'http://13.209.12.14/api';
