+++ src/Constants/constants.ts
@@ -1,0 +1,14 @@
+export const TOPIC_CATEGORY = [
+  {
+    text: '엔터테인먼트',
+    key: 'ENTERTAINMENT',
+  },
+  {
+    text: '생활/취미',
+    key: 'LIFE_HOBBY',
+  },
+  {
+    text: '지식/기타',
+    key: 'KNOWLEDGE_ETC',
+  },
+];
