+++ src/Modals/ModalBlur.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {StyleSheet, View} from 'react-native';
-import {SCREEN_HEIGHT, SCREEN_WIDTH} from '../constants';
+import {SCREEN_HEIGHT, SCREEN_WIDTH} from '../Constants/screen';
 
 export function ModalBlur() {
   return <View style={styles.modalBlur} />;
